split NoQuiz and NoQuizMain for 가독성성
